requirements.txt
streamlit
numpy
plotly
pandas
python-dotenv
langchain
langchain-community
openai
pypdf